feat(app): 1592 - factorized apply button
